Reverted session id from guid to string of length 16 or 32 bytes. Made window id have length of 4 or 8 bytes. Fixed problem with receiving window id 0 at almost any time, by requiring window id wrap around at 9E18. Added large bandwidth-delay product feature of protocol to README. Removed need for opening of session to indicate PDU to send.

+++ csharp/ScalableIPC.Core/Session/SendHandlerAssistant.cs
@@ -49,7 +49,7 @@ namespace ScalableIPC.Core.Session
         private void ContinueSending()
         {
             var nextMessage = CurrentWindow[_sentPduCount];
-            int windowIdSnapshot = _sessionHandler.NextWindowIdToSend;
+            long windowIdSnapshot = _sessionHandler.NextWindowIdToSend;
             nextMessage.WindowId = windowIdSnapshot;
             nextMessage.SequenceNumber = _sentPduCount - PreviousSendCount;
             _sessionHandler.EndpointHandler.HandleSend(_sessionHandler.RemoteEndpoint, nextMessage)
@@ -112,7 +112,7 @@ namespace ScalableIPC.Core.Session
             }
         }
 
-        private AbstractPromise<VoidType> HandleSendSuccess(int windowIdSnapshot)
+        private AbstractPromise<VoidType> HandleSendSuccess(long windowIdSnapshot)
         {
             _sessionHandler.EventLoop.PostCallback(() =>
             {
